card的test_code動態增列+save array
