+++ app/javascript/stylesheets/application.scss
@@ -4,4 +4,4 @@
 @import "components/buttons";
 @import "components/forms";
 @tailwind utilities;
-@import "pages/editor";
+@import "pages/cards";
