+++ app/javascript/stylesheets/pages/_cards.scss
@@ -1,3 +1,3 @@
 .editor {
     min-height: 6rem;
-}
+} //
